refactor: move playbackReducer to it's own file
